More GUI tweaks; Deformation code cleanup
Former-commit-id: b4ccf6570eeb0112aad146268c82ab0283cf0b5e
Former-commit-id: 9e38dea1cbeb3cec76a480c06a697d41059263b6

+++ src/ofApp.h
@@ -194,7 +194,7 @@ public:
 
     // timeline
     ofxTimeline timeline;
-    float timelineDurationSeconds;
+    int timelineDurationSeconds;
     float timelinePreviousDuration;
     void timelineSetup(float duration);
     void timelineUpdate();
